Move league stats tests to Stat Tracker test
